# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "The one and only error shape returned by every 4xx/5xx response.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "Farm": {
        "description": "Full server representation of a registered farm.",
        "properties": {
          "area_hectares": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              12.5
            ],
            "title": "Area Hectares"
          },
          "country_code": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 3166-1 alpha-2, uppercase.",
            "examples": [
              "KE"
            ],
            "title": "Country Code"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "crop_count": {
            "default": 0,
            "description": "Number of crops currently attached to this farm.",
            "minimum": 0.0,
            "title": "Crop Count",
            "type": "integer"
          },
          "elevation_m": {
            "anyOf": [
              {
                "maximum": 9000.0,
                "minimum": -500.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "farming_practice": {
            "$ref": "#/components/schemas/FarmingPractice",
            "default": "conventional"
          },
          "has_analysis": {
            "default": false,
            "description": "Whether an analysis run exists. Drives the dashboard empty state.",
            "title": "Has Analysis",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "irrigation_type": {
            "$ref": "#/components/schemas/IrrigationType",
            "default": "rainfed"
          },
          "latitude": {
            "examples": [
              -1.2921
            ],
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "examples": [
              36.8219
            ],
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "examples": [
              "North Field"
            ],
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "region": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Nairobi"
            ],
            "title": "Region"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "Farm",
        "type": "object"
      },
      "FarmCreate": {
        "description": "Request body for `POST /api/v1/farms`.",
        "properties": {
          "area_hectares": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              12.5
            ],
            "title": "Area Hectares"
          },
          "country_code": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 3166-1 alpha-2, uppercase.",
            "examples": [
              "KE"
            ],
            "title": "Country Code"
          },
          "elevation_m": {
            "anyOf": [
              {
                "maximum": 9000.0,
                "minimum": -500.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "farming_practice": {
            "$ref": "#/components/schemas/FarmingPractice",
            "default": "conventional"
          },
          "irrigation_type": {
            "$ref": "#/components/schemas/IrrigationType",
            "default": "rainfed"
          },
          "latitude": {
            "examples": [
              -1.2921
            ],
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "examples": [
              36.8219
            ],
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "examples": [
              "North Field"
            ],
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "region": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Nairobi"
            ],
            "title": "Region"
          }
        },
        "required": [
          "name",
          "latitude",
          "longitude"
        ],
        "title": "FarmCreate",
        "type": "object"
      },
      "FarmCrop": {
        "description": "A crop planted on a specific farm, with the catalog entry embedded so the\nfrontend renders a card without a second request.",
        "properties": {
          "area_hectares": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Hectares"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "crop": {
            "$ref": "#/components/schemas/Crop"
          },
          "crop_id": {
            "description": "References a crop in the reference catalog.",
            "format": "uuid",
            "title": "Crop Id",
            "type": "string"
          },
          "expected_harvest_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Harvest Date"
          },
          "farm_id": {
            "format": "uuid",
            "title": "Farm Id",
            "type": "string"
          },
          "growth_stage": {
            "$ref": "#/components/schemas/GrowthStage",
            "default": "not_planted"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_primary": {
            "default": false,
            "description": "The crop this farm's analysis is centred on. At most one per farm.",
            "title": "Is Primary",
            "type": "boolean"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "planting_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planting Date"
          },
          "status": {
            "$ref": "#/components/schemas/CropStatus",
            "default": "planned"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "crop_id",
          "id",
          "farm_id",
          "crop",
          "created_at",
          "updated_at"
        ],
        "title": "FarmCrop",
        "type": "object"
      },
      "FarmCropCreate": {
        "description": "Request body for `POST /api/v1/farms/{farm_id}/crops`.",
        "properties": {
          "area_hectares": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Hectares"
          },
          "crop_id": {
            "description": "References a crop in the reference catalog.",
            "format": "uuid",
            "title": "Crop Id",
            "type": "string"
          },
          "expected_harvest_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Harvest Date"
          },
          "growth_stage": {
            "$ref": "#/components/schemas/GrowthStage",
            "default": "not_planted"
          },
          "is_primary": {
            "default": false,
            "description": "The crop this farm's analysis is centred on. At most one per farm.",
            "title": "Is Primary",
            "type": "boolean"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "planting_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planting Date"
          },
          "status": {
            "$ref": "#/components/schemas/CropStatus",
            "default": "planned"
          }
        },
        "required": [
          "crop_id"
        ],
        "title": "FarmCropCreate",
        "type": "object"
      },
      "FarmCropList": {
        "description": "Response for `GET /api/v1/farms/{farm_id}/crops`.",
        "properties": {
          "has_next": {
            "default": false,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FarmCrop"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total items matching the query.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "FarmCropList",
        "type": "object"
      },
...
      "FarmList": {
        "description": "Response for `GET /api/v1/farms`.",
        "properties": {
          "has_next": {
            "default": false,
            "title": "Has Next",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Farm"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 50,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "description": "Total items matching the query.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "FarmList",
        "type": "object"
      },
      "FarmUpdate": {
        "description": "Request body for `PATCH /api/v1/farms/{farm_id}`. Omitted fields are unchanged.",
        "properties": {
          "area_hectares": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area Hectares"
          },
          "country_code": {
            "anyOf": [
              {
                "maxLength": 2,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "elevation_m": {
            "anyOf": [
              {
                "maximum": 9000.0,
                "minimum": -500.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elevation M"
          },
          "farming_practice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FarmingPractice"
              },
              {
                "type": "null"
              }
            ]
          },
          "irrigation_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IrrigationType"
              },
              {
                "type": "null"
              }
            ]
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "region": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          }
        },
        "title": "FarmUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\n**TerraNex** is the central intelligence layer for agricultural decision support.\n\nIt fuses farm, crop, soil, weather and vegetation data with AI reasoning to produce\nfarm health analysis, weather and water risk, disease risk, agricultural advisories,\ncrop and regenerative-agriculture recommendations, and multimodal crop-image diagnosis.\n\n### For frontend developers\n\n* Every endpoint lives under `/api/v1`. Field casing is `snake_case` throughout.\n* Every failure returns the same envelope:\n  `{\"error\": {\"code\", \"message\", \"details\", \"request_id\"}}`.\n  **Branch on `code`** \u2014 `message` is for humans and may change.\n* Every response carries `X-Request-Id`; include it when reporting a bug.\n* Generate typed bindings instead of hand-writing them:\n  `npx openapi-typescript ../contracts/openapi.json -o src/api/types.gen.ts`\n",
    "title": "TerraNex API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/farms": {
      "get": {
        "operationId": "list_farms_api_v1_farms_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "List farms",
        "tags": [
          "farms"
        ]
      },
      "post": {
        "description": "Creates a farm at the given coordinates. Coordinates drive every downstream data lookup.",
        "operationId": "create_farm_api_v1_farms_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Farm"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Register a farm",
        "tags": [
          "farms"
        ]
      }
    },
    "/api/v1/farms/{farm_id}": {
      "delete": {
        "description": "Soft delete. Historical analysis runs are retained.",
        "operationId": "delete_farm_api_v1_farms__farm_id__delete",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Delete a farm",
        "tags": [
          "farms"
        ]
      },
      "get": {
        "operationId": "get_farm_api_v1_farms__farm_id__get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Farm"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Get a farm",
        "tags": [
          "farms"
        ]
      },
      "patch": {
        "description": "Partial update. Omitted fields are left unchanged.",
        "operationId": "update_farm_api_v1_farms__farm_id__patch",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Farm"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Update a farm",
        "tags": [
          "farms"
        ]
      }
    },
...
    "/api/v1/farms/{farm_id}/crops": {
      "get": {
        "operationId": "list_farm_crops_api_v1_farms__farm_id__crops_get",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmCropList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "List a farm's crops",
        "tags": [
          "farms"
        ]
      },
      "post": {
        "description": "Attaches a catalog crop to this farm with planting dates and growth stage.",
        "operationId": "add_farm_crop_api_v1_farms__farm_id__crops_post",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmCropCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmCrop"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or CROP_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Add a crop to a farm",
        "tags": [
          "farms"
        ]
      }
    },
    "/api/v1/farms/{farm_id}/crops/{farm_crop_id}": {
      "delete": {
        "operationId": "delete_farm_crop_api_v1_farms__farm_id__crops__farm_crop_id__delete",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "description": "Farm-crop (planting) identifier.",
            "in": "path",
            "name": "farm_crop_id",
            "required": true,
            "schema": {
              "description": "Farm-crop (planting) identifier.",
              "format": "uuid",
              "title": "Farm Crop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or CROP_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Remove a crop from a farm",
        "tags": [
          "farms"
        ]
      },
      "patch": {
        "description": "Used to advance growth stage or record the harvest.",
        "operationId": "update_farm_crop_api_v1_farms__farm_id__crops__farm_crop_id__patch",
        "parameters": [
          {
            "description": "Farm identifier.",
            "in": "path",
            "name": "farm_id",
            "required": true,
            "schema": {
              "description": "Farm identifier.",
              "format": "uuid",
              "title": "Farm Id",
              "type": "string"
            }
          },
          {
            "description": "Farm-crop (planting) identifier.",
            "in": "path",
            "name": "farm_crop_id",
            "required": true,
            "schema": {
              "description": "Farm-crop (planting) identifier.",
              "format": "uuid",
              "title": "Farm Crop Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FarmCropUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FarmCrop"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "FARM_NOT_FOUND or CROP_NOT_FOUND"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope"
          }
        },
        "summary": "Update a farm's crop",
        "tags": [
          "farms"
        ]
      }
    },
...
  }
}